third_party/WebKit/LayoutTests/svg/hixie/error/013-expected.txt

CONSOLE ERROR: line 10: Error: <svg> attribute width: A negative value is not valid. ("-100")
